Fix bug in TestSendDisembargo/SendQueuedResultToCaller
Sadly, I don't think this is #426, but I noticed it while hunting that
down: we set Which to receiverHosted but then set the SenderHosted
field. This patch fixes that.

I think we're getting lucky in the test such that it passes anyway
because we only ever put one export on the wire so the ID is always zero
anyway.