Use go-util module, with some trivial substitutions
Per discussion on matrix, we have a fair bit of random utility code, and
separately I've been accumulating a repo of stuff I've been using in
tempest. Let's use it here too.

There are more things we could replace, as well as some packages that
should get moved *into* go-util, but this is a start.