Avoid use of fmt in the main package.
This adds a String() method to ClientHook so we can avoid using fmt for
that, then removes the last use of fmt in the main package, per #364.

Before closing that issue, I think I'd still like to get it out of the
flowcontrol package hierarchy at least, and maybe elsewhere if
practical.