Don't register schema by default.
This patch changes code generation so that it doesn't register schema
with the schemas package's DefaultRegistry by default. Instead, we
expose a RegisterSchemas function which must be invoked on the relevant
registry.

This is a step towards #89. The default registry still exists, and APIs
like pogs still use it implicitly, so there is more work to do.

This should make it a bit easier for the compiler to do dead-code
elimination for the schema blobs. This shaves about 28K off the tempest
WASM blob. (432K down from 460K).