Add a test for disembargos on senderPromises.
This caught a bug: we need to check if ClientState.Metadata is nil,
which is possible if the client itself is null, and seems to happen to
the promise after it is resolved.

TODO: we should de-dup the logic between releaseExport and
releaseExports. This is not entirely trivial though, because the
latter is executed after we've wiped the exports table.