Correctly send receiverLoopbacks that target promisedAnswers.
Worryingly, this was manifesting in the test as a deadlock: we hit the
error complaining about it not being an import, but then connection
shutdown hung, waiting on tasks. I haven't pinned down exactly what was
going on there, but this sidesteps the issue by fixing the thing that
was causing a connection abort in the first place.