Get rid of lockedConn.assertIs().
This is obsolete, now that we have to go through c.lk to get to the
entry.

This patch also drops all of the parameters that were transitively
passed to assertIs().

We do still sometimes want to actually *use* the connection though,
so we add a method to ansent to fetch it, which localizes the cast.