Fix #348
This Shutdown() is redundant with the one in clientPromise.shutdown().
When #348 was filed, this was a close() on the done channel instead of a
call to ClientHook.Shutdown, but became the latter after merging
calls and refs -- done was for the former, Shutdown() for the latter.

I'm still not 100% sure why this used to be non-deterministic.