Avoid object cycle in clientHooks.
Currently, non-promises are represented by clientHooks whose
resolevedHook points back the same clientHook.

I want to move these over to the rc package from go-util, but having
object cycles like this is going to cause problems. Instead, we just
keep track of whether this was a promise in the first place more
explicitly: wrap resolvedHook in a Maybe, where absent means this was
never a promise.

We could use a boolean flag here, but this is more type safe.