promiseState: remove obsolete ongoingCalls and callsStopped fields.
This patch has no functional impact; we never actually do a receive on
callsStopped. There is a now-removed receive on a local callsStopped
variable, but it is never assigend -- it was obviously supposed to be
at some point, but I believe we want to remove this anyway, since we are
in general no longer blocking on calls before shutting down.