Delete irrelevant FIXME comment.
I noticed that there's nowhere where we actually hold locks on multiple
Promise.states. So there's no need to even have a rule about this. I'd
added the FIXME at some point when reading the comment about the
invariant because it didn't seem right -- but it turns out it doesn't
matter.