rpc: Add extra releasers to Client
This commit adds a new feature: additional releaser functions tracked by
a Client instance that are called when that client itself is released.

These releasers can be added by application code or by library code to
ensure certain cleanup procedures only happen when the client is
explicitly marked as released.