rpc: Release Bootstrap promise clients only on shutdown
This changes the Bootstrap function of RPC connections to release the
promise's clients only when the external client is released instead of
immediately after the promise is fulfilled.

This fixes an issue where application code may still be in the process
of using this client when it is released and nilled when the
corresponding Return message is received and the goroutine of the
existing implementation is unblocked inside the ReleaseClients call.