main: Fix buffer reuse read benchmark
Previously, this benchmark was using the Reset() method to reset the
message's arena, which can't be done for reading, as the recently
introduced comments indicate.

This fixes the benchmark by releasing the previous arena and directly
modifying the arena field of the message.

In the future, this should be done explicitly by either Reset() or by a
new dedicated function.