transport: Fix double release of incoming message
This fixes the double release panic introduced in the prior commit. The
panic was only introduced to prove that double release is possible happening.

Note: In my (matheusd) opinion, Release() calls should panic or error in
case of double release so that callers are fixed and they themselves do
not rely on this protection. This is suggested as a future change
because it potentially breaks clients that were relying on this
protection.