message: Make NewMessage usable for reading
Previously, NewMessage() would always attempt to allocate the first
word of the first segment and would consider arenas with more than 1
segment as not usable.

This restriction is lifted in this commit by only attempting to allocate
the first segment and not the root pointer space. This allows passing a
fully populated Arena object to Reset(), which in turns allows passing
it to NewMessage.

This will be helpful by ensuring NewMessage can be used in all
situations where a message is needed (not just for writing).

NOTE: this is a breaking API change. Before this commit, NewMessage
would fail for read-only arenas or that otherwise could not allocate
space for the root pointer.

The allocation for the root pointer space is now enforced only when
(1) performing an initial allocation or (2) setting the root pointer.

Enforcement on (2) is only necessary because one call path in file
fileparts.go makes a call to SetRoot before attempting to allocate space
in the message. This is likely an erroneous call and should be
investigated in the future in order to simplify the code.