bufferpool: Improve performance
This improves the performance of bufferpool by removing the mutex called
on every operation (via init()) and improving bucket indexing by
removing the loop and using direct index calculation.

Warning: this is a breaking API change. Zero valued Pool objects are no
longer safe to be used and MUST be initialized by NewPool().