MultiSegmentArena: Fix release back into bufferpool
This fixes a bug that was preventing correct release of buffers
allocated by Decode() back to the global buffer pool.

Previously, the data variable was overriden on the loop that splits the
segments, preventing the correct tracking of the full data buffer and
its subsequent return back to the buffer pool when the arena was
released.

This also adds a .Release() call to the Decode benchmark to better
reflect the ability to reuse buffers.