Fix broken AnswerQueue behavior with chained pipeline calls
AnswerQueue does not handle chained pipeline calls correctly.
I'm assuming this is because the code for it might be incomplete. It doesn't assign basis recv functions for any bases other than 0 and uses the wrong value for the basis itself on subsequent chained calls.

This fixes those bugs and adds more documentation about how AnswerQueue is meant to function because I had to stare at it for 2 hours to figure it out.